[refactor] 팀 논의 반영: AuthUtil을 컴포넌트로 변경, DI 받아서 다형성 갖게 함
