+++ src/main/java/com/prgrms/ohouse/domain/common/security/AuthUtility.java
@@ -1,0 +1,7 @@
+package com.prgrms.ohouse.domain.common.security;
+
+import com.prgrms.ohouse.domain.user.model.User;
+
+public interface AuthUtility {
+	User getAuthUser();
+}
